+++ linking-platform/link-picker/src/index.ts
@@ -1,1 +1,0 @@
-export { LinkPicker } from './link-picker';