+++ linking-platform/link-picker/src/link-picker.tsx
@@ -1,3 +1,0 @@
-export const LinkPicker = () => {
-  return null;
-};