+++ media/smart-card/src/view/FlexibleCard/components/actions/action/__tests__/deleteAction.test.tsx
@@ -9,33 +9,48 @@ import CrossIcon from '@atlaskit/icon/glyph/cross';
 describe('Action: Delete', () => {
   const testId = 'smart-action-delete-action';
 
-  it('should render DeleteAction with some text', async () => {
-    const text = 'spaghetti';
-    const onClick = () => {};
-    const { getByTestId } = render(
-      <DeleteAction onClick={onClick} content={text} testId={testId} />,
-    );
+  describe.each([
+    ['as dropdown item', true],
+    ['as button', false],
+  ])('%s', (_: string, asDropdownItem: boolean) => {
+    it('should render DeleteAction with some text', async () => {
+      const text = 'spaghetti';
+      const onClick = () => {};
+      const { getByTestId } = render(
+        <DeleteAction
+          asDropDownItem={asDropdownItem}
+          onClick={onClick}
+          content={text}
+          testId={testId}
+        />,
+      );
 
-    const element = await waitForElement(() => getByTestId(testId));
+      const element = await waitForElement(() => getByTestId(testId));
 
-    expect(element).toBeTruthy();
-    expect(element.textContent).toBe('spaghetti');
-  });
+      expect(element).toBeTruthy();
+      expect(element.textContent).toBe('spaghetti');
+    });
 
-  it('should call the supplied onClick when button is clicked', async () => {
-    const text = 'spaghetti';
-    const mockOnClick = jest.fn();
-    const { getByTestId } = render(
-      <DeleteAction onClick={mockOnClick} content={text} testId={testId} />,
-    );
+    it('should call the supplied onClick when button is clicked', async () => {
+      const text = 'spaghetti';
+      const mockOnClick = jest.fn();
+      const { getByTestId } = render(
+        <DeleteAction
+          asDropDownItem={asDropdownItem}
+          onClick={mockOnClick}
+          content={text}
+          testId={testId}
+        />,
+      );
 
-    const element = await waitForElement(() => getByTestId(testId));
+      const element = await waitForElement(() => getByTestId(testId));
 
-    expect(element).toBeTruthy();
-    expect(element.textContent).toBe('spaghetti');
+      expect(element).toBeTruthy();
+      expect(element.textContent).toBe('spaghetti');
 
-    userEvent.click(element);
-    expect(mockOnClick).toHaveBeenCalled();
+      userEvent.click(element);
+      expect(mockOnClick).toHaveBeenCalled();
+    });
   });
 
   describe('size', () => {