+++ editor/embedded-document/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @atlaskit/embedded-document
 
+## 0.7.40
+
+### Patch Changes
+
+- Updated dependencies
+
+## 0.7.39
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 0.7.38
 
 ### Patch Changes